Separated interfaces and models
