+++ src/graphql/queries/RootQuery.ts
@@ -4,7 +4,7 @@ import { GraphQLBoolean, GraphQLInt, GraphQLList, GraphQLObjectType, GraphQLStri
 // TYPES
 import { User } from '../types/UserTypes';
 import { Tweet, TweetList } from '../types/TweetTypes';
-import { TweetFilter } from '../../types/args/TweetFilter';
+import { TweetFilter } from '../../models/args/TweetFilter';
 
 // RESOLVERS
 import UserResolver from '../resolvers/UserResolver';
