+++ src/models/data/CursoredData.ts
@@ -1,10 +1,5 @@
-// SERVICES
-import { AccountService } from "../../services/accounts/AccountService";
-import { TweetService } from "../../services/data/TweetService";
-import { UserService } from "../../services/data/UserService";
-
-// INTERFACES
-import { CursoredDataInterface } from "../interfaces/Service";
+// TYPES
+import { CursoredDataInterface } from "../../types/interfaces/Service";
 
 /**
  * The cursor to the batch of data to be fetched.
@@ -46,20 +41,4 @@ export class CursoredData<T> implements CursoredDataInterface<T> {
         this.list = list;
         this.next = new Cursor(next);
     }
-}
-
-/**
- * The data context from where data is to be fetched.
- * 
- * @public
- */
-export interface DataContext {
-    /** Handles data related to users. */
-    users: UserService,
-
-    /** Handles data related to tweets. */
-    tweets: TweetService,
-
-    /** Handles account related operations. */
-    account: AccountService
 }
