+++ src/types/interfaces/Resolvers.ts
